ARUHA-855 Remove unneeded code
