+++ src/main/java/org/zalando/nakadi/service/AuthorizationValidator.java
@@ -2,6 +2,15 @@ package org.zalando.nakadi.service;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableMap;
+import static com.google.common.collect.Lists.newArrayList;
+import static com.google.common.collect.Sets.newHashSet;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.stream.Collectors;
+import javax.annotation.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.zalando.nakadi.domain.EventType;
@@ -9,6 +18,7 @@ import org.zalando.nakadi.domain.EventTypeAuthorization;
 import org.zalando.nakadi.domain.EventTypeResource;
 import org.zalando.nakadi.exceptions.ForbiddenAccessException;
 import org.zalando.nakadi.exceptions.InvalidEventTypeException;
+import org.zalando.nakadi.exceptions.ResourceAccessNotAuthorizedException;
 import org.zalando.nakadi.exceptions.ServiceUnavailableException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
 import org.zalando.nakadi.plugin.api.PluginException;
@@ -47,6 +57,21 @@ public class AuthorizationValidator {
                     "writers", auth.getWriters());
             checkAuthAttributesAreValid(allAttributes);
             checkAuthAttributesNoDuplicates(allAttributes);
+        }
+    }
+
+    public void validateWrite(final EventType eventType) {
+        if (eventType.getAuthorization() == null) {
+            return;
+        }
+        final MyEventTypeResource2 resource = new MyEventTypeResource2(
+                eventType.getName(), eventType.getAuthorization());
+        final boolean authorized = authorizationService.isAuthorized(
+                null,
+                AuthorizationService.Operation.WRITE,
+                resource);
+        if (!authorized) {
+            throw new ResourceAccessNotAuthorizedException(AuthorizationService.Operation.WRITE, resource);
         }
     }
