+++ src/test/java/org/zalando/nakadi/service/EventPublisherTest.java
@@ -41,7 +41,6 @@ import org.zalando.nakadi.exceptions.IllegalScopeException;
 import org.zalando.nakadi.exceptions.PartitioningException;
 import org.zalando.nakadi.exceptions.ResourceAccessNotAuthorizedException;
 import org.zalando.nakadi.partitioning.PartitionResolver;
-import org.zalando.nakadi.plugin.api.authz.AuthorizationService;
 import org.zalando.nakadi.repository.TopicRepository;
 import org.zalando.nakadi.repository.db.EventTypeCache;
 import org.zalando.nakadi.security.Client;
@@ -77,7 +76,7 @@ public class EventPublisherTest {
     private final PartitionResolver partitionResolver = mock(PartitionResolver.class);
     private final TimelineSync timelineSync = mock(TimelineSync.class);
     private final Enrichment enrichment = mock(Enrichment.class);
-    private final AuthzChecker authzChecker = mock(AuthzChecker.class);
+    private final AuthorizationValidator authzValidator = mock(AuthorizationValidator.class);
     private final NakadiSettings nakadiSettings = new NakadiSettings(0, 0, 0, TOPIC_RETENTION_TIME_MS, 0, 60,
             NAKADI_POLL_TIMEOUT, NAKADI_SEND_TIMEOUT, TIMELINE_WAIT_TIMEOUT_MS, NAKADI_EVENT_MAX_BYTES,
             NAKADI_SUBSCRIPTION_MAX_PARTITIONS);
@@ -91,7 +90,7 @@ public class EventPublisherTest {
         Mockito.when(ts.getTimeline(any())).thenReturn(timeline);
 
         publisher = new EventPublisher(ts, cache, partitionResolver, enrichment, nakadiSettings, timelineSync,
-                authzChecker);
+                authzValidator);
     }
 
     @Test
@@ -114,11 +113,8 @@ public class EventPublisherTest {
         mockSuccessfulValidation(et);
 
         Mockito.doThrow(new ResourceAccessNotAuthorizedException(null, null))
-                .when(authzChecker)
-                .performCheck(
-                        Mockito.eq(et),
-                        Mockito.eq(AuthorizationService.Operation.WRITE),
-                        Mockito.any());
+                .when(authzValidator)
+                .validateWrite(Mockito.eq(et));
 
         publisher.publish(buildDefaultBatch(1).toString(), et.getName(), FULL_ACCESS_CLIENT);
     }
